refactor(cli): switch legacy analytics to new telemetry
