+++ packages/@expo/cli/src/api/user/user.ts
@@ -5,9 +5,9 @@ import UserSettings from './UserSettings';
 import { getSessionUsingBrowserAuthFlowAsync } from './expoSsoLauncher';
 import { CurrentUserQuery } from '../../graphql/generated';
 import * as Log from '../../log';
-import * as Analytics from '../../utils/analytics/rudderstackClient';
 import { getDevelopmentCodeSigningDirectory } from '../../utils/codesigning';
 import { env } from '../../utils/env';
+import { telemetry } from '../../utils/telemetry';
 import { getExpoWebsiteBaseUrl } from '../endpoint';
 import { graphqlClient } from '../graphql/client';
 import { UserQuery } from '../graphql/queries/UserQuery';
@@ -42,13 +42,7 @@ export async function getUserAsync(): Promise<Actor | undefined> {
   if (!env.EXPO_OFFLINE && !currentUser && hasCredentials) {
     const user = await UserQuery.currentUserAsync();
     currentUser = user ?? undefined;
-    if (user) {
-      await Analytics.setUserDataAsync(user.id, {
-        username: getActorDisplayName(user),
-        user_id: user.id,
-        user_type: user.__typename,
-      });
-    }
+    telemetry?.identify(currentUser);
   }
   return currentUser;
 }
