+++ packages/@expo/cli/src/start/server/DevServerManager.ts
@@ -7,8 +7,8 @@ import DevToolsPluginManager from './DevToolsPluginManager';
 import { getPlatformBundlers } from './platformBundlers';
 import { Log } from '../../log';
 import { FileNotifier } from '../../utils/FileNotifier';
-import { logEventAsync } from '../../utils/analytics/rudderstackClient';
 import { env } from '../../utils/env';
+import { logEventAsync } from '../../utils/telemetry';
 import { ProjectPrerequisite } from '../doctor/Prerequisite';
 import { TypeScriptProjectPrerequisite } from '../doctor/typescript/TypeScriptProjectPrerequisite';
 import { printItem } from '../interface/commandsTable';
